+++ packages/import/import-flat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @porulle/import-flat
 
+## 0.10.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @porulle/core@0.10.6
+
 ## 0.10.5
 
 ### Patch Changes