+++ packages/plugins/plugin-pos-restaurant/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @porulle/plugin-pos-restaurant
 
+## 0.10.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @porulle/core@0.10.6
+  - @porulle/plugin-pos@0.10.6
+
 ## 0.10.5
 
 ### Patch Changes